seperate code for integrators and particles classes
